+++ src/loaders/components/path-form.tsx
@@ -9,22 +9,22 @@ export default function FolderPathFormLoader({items = 5}: Props){
           <div className="border border-accent animate-pulse flex flex-col gap-4 rounded-xl p-6">
                <div className="flex justify-between items-center gap-2 w-full">
                     <div className="space-y-2 w-full">
-                         <div className="h-4 bg-accent rounded-md w-1/3"/>
-                         <div className="h-3.5 bg-accent rounded-md w-1/5"/>
+                         <div className="h-4 bg-muted rounded-md w-1/3"/>
+                         <div className="h-3.5 bg-muted rounded-md w-1/5"/>
                     </div>
-                    <div className="h-9 w-40 bg-accent rounded-md"/>
+                    <div className="h-9 w-40 bg-muted rounded-md"/>
                </div>
                {range.length>0 ? (
                     <div className="space-y-2">
                          {range.map(id=>(
                               <div key={id} className="flex justify-between items-center gap-2 pb-1 border-b last:pb-0 last:border-none">
-                                   <div className="bg-accent rounded-md w-1/2 h-4"/>
-                                   <div className="bg-accent rounded-md size-10"/>
+                                   <div className="bg-muted rounded-md w-1/2 h-4"/>
+                                   <div className="bg-muted rounded-md size-10"/>
                               </div>
                          ))}
                     </div>
                ) : (
-                    <div className="bg-accent rounded-md w-full h-4"/>
+                    <div className="bg-muted rounded-md w-full h-4"/>
                )}
           </div>
      )