+++ src/loaders/components/settings-item.tsx
@@ -6,9 +6,9 @@ interface Props{
 export default function SettingsItemLoader({className, children, noDescription=false}: Props){
      return (
           <div className="border border-accent animate-pulse flex flex-col gap-6 rounded-xl p-6">
-               <div className="h-4 bg-accent rounded-md w-1/3"/>
+               <div className="h-4 bg-muted rounded-md w-1/3"/>
                {!noDescription && (
-                    <div className="h-3.5 bg-accent rounded-md w-1/2"/>
+                    <div className="h-3.5 bg-muted rounded-md w-1/2"/>
                )}
                {!className ? children : (
                     <div className={className}>