+++ CHANGELOG.md
@@ -18,6 +18,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
   - Real-Time Scan
   - Themes
 - Memoized functions inside contexts and on definitions updater state to reduce re-rendering
+- Skeleton Loaders - Change the background image to a muted color
 ### Fixed
 - macOS Homebrew ClamAV detection fallback - Fixes [#16](https://github.com/ArsenTech/clamav-gui/issues/16)
   - `/opt/homebrew/bin/clamscan`